Add tests for `€` characters in identifiers.
The `€` character is a valid character in a Java identifier, so it
stays as is in the `Name`s of the IR. However, it is not a valid
JavaScript identifier part, so it needs to be encoded by the
emitter.

This was fixed by b8acd0349f264303763abb1178d8a04d49937c62 but no
test for this behavior was added at the time.