Fix #3901: Do not fill `groupStartMap` for non-matching groups.
`groupStartMap` is initially filled with `-1` everywhere, which is
the correct value for non-matching groups. When propagating the
`start` and `end` of a `GroupNode`, we previously always overrode
that value with `start`, which can `>= 0` even for non-matching
groups if they are within a negative look-ahead. This is incorrect,
as non-matching groups must always report `-1` as their `start`.

We now explicitly avoid overriding the initial `-1` for
non-matching groups. This more closely corresponds to what was done
in `GroupStartMap.setEndReturnStart` and `setStartReturnEnd` before
the big rewrite of 56dd7abb9c32ae519c8327ee574295687ca96127.