[no-master] Disable testing the test suite with PhantomJS on the CI.
PhantomJS is still by far the biggest cause of spurious failures of
the CI. Since I have not seen any genuine bug uncovered by testing
on PhantomJS (only) in the last few *years*, the cost of this test
overwhelmingly exceeds its potential benefits.

Therefore, in this commit we simply disable the last command that
was running the test suite with PhantomJS, as well as the last
command that was running non-optimized code with PhantomJS.

PhantomJS is still minimally tested by its unit tests, as well as
in the helloworld and the testingExample (in standard fastOpt and
fullOpt modes).