Fix #3933: Use fresh names everywhere in GenJSExports.
The logic in `GenJSExports.scala` was always using
compiler-generated local names that were not fresh. This was fine
before, because the local scope of an exporter method was never
mixed with that of user-defined code of even scalac-generated code.
However, it became problematic with the introduction of
Scala.js-defined classes, because their constructors precisely
create such mixes.

We now use proper fresh local names everywhere when creating
exporter methods.