Respect sbt `repositories` when fetching the linker
sbt has a ~/.sbt/repositories file that can be used to add extra default
resolvers, this is especially useful for setting up a cache or a proxy
(see https://www.scala-sbt.org/1.x/docs/Proxy-Repositories.html for more
information).

So far, sbt-scalajs did not respect this setting and instead manually
specified a list of resolvers, this commit fixes this. Ideally we would
just use `dependencyResolution.value` but sbt does not define this task
in the global scope.