Make CacheBox public.
It is used in the types of public settings, so it must be public.

Since we make it public, we must also make sure that it works even
if `T` is nullable and that `null` is a valid value. So we change
the implementation not to rely on `null` as a sentinel.