Fix #3948: Use sbt's `dependencyResolution` for the LinkerImpl.
Instead of always creating a default `IvyDependencyResolution` to
resolve the scalajs-linker artifact, used to create the
`LinkerImpl`, we now use sbt's built-in `dependencyResolution`
task.

However, since that task is only available per project (and not in
the Global scope), we fetch it from the `LocalRootProject`. In the
rare case that the root project disabled the built-in `IvyPlugin`,
we fall back to the default `IvyDependencyResolution` as before.

With these changes, resolution of the scalajs-linker artifact will
follow custom configurations, such as the `~/.sbt/repositories`
file.