Extract the classpath used for LinkerImpl in a separate key.
We use `fullClasspath in scalaJSLinkerImpl` as another hook for
customization of the linker impl, which does not require users to
deal with the caching.

In addition, we rename `LinkerImpl.default` to `reflect`, as it is
clearer what it does, and it might not be used so much by clients
anymore.