Remove the override of `Throwable` in the minilib.
The only blocker was calling `Utils.isUndefined`, which we manually
inline instead.

The minilib is now free of any override. It is strictly a subset of
the standard java lib. Therefore, it does not need its own project
anymore, and is turned into just a custom task of the library.