Optimize `x.getClass().getName()` not to instantiate `j.l.Class`.
There are a number of use cases for `x.getClass().getName()` in
core code, including in `Object.toString()`, `Throwable.toString()`
and `MatchError.getMessage()`. These normally require instantiating
a `java.lang.Class` from its class data, then to fetch the name
from the data.

We make this more efficient by intrinsifying `Class.getName()` with
a more direct implementation if the receiver is a `GetClass(x)` or
a `ClassOf(C)`.

With this optimization, it becomes possible for a codebase not to
have any `ClassOf` or `GetClass` node at all, and hence for
`java.lang.Class` not to be needed at all. In that case, the core
JS lib was leaving an erroneous reference to it. To solve this, and
optimize a bit more in the process, we add checks to remove
`j.l.Class` support from the core JS lib if `j.l.Class` is never
instantiated.

The above scenario actually happens for the hello world example.