Remove the override for RedBlackTree.scala in 2.12.2+.
It hasn't been necessary since 2.12.2, because our fix was merged
upstream in that version.

This commit is an update of 36ea85c34e647a18894154df852ac774976daada
which had already removed the override for 2.13.x.

This change is necessary now because 2.12.12+ has a completely
different implementation, backported from 2.13.x in the upstream
pull requests https://github.com/scala/scala/pull/8749, which is
not compatible with the classes that use `RedBlackTree`.