Fix #4021: Filter fields in Refiner exactly like BaseLinker.
Previously, the Refiner would eliminate *all* fields in
non-instantiated classes. This would remove static fields even if
they are needed, which causes invalid .js files to be generated.

The BaseLinker has been doing the right thing since the commit
9328d629294defc6eb75a7ffd0907068ce02f984, but the Refiner had not
been updated. We now make them both do the same thing, which solves
the issue.

Since the issue is not reproducible with Scala.js source code, we
test this with a linker test.