Fix #3500: Implement @js.native vals and defs.
We now allow `val`s and `def`s (but not `lazy val`s, `var`s or
setter `def`s) in Scala static objects to be annotated with
`@js.native`. For example:

  object QueryStringWithNativeDef {
    @js.native
    @JSImport("querystring", "stringify")
    def stringify(obj: js.Dictionary[String], sep: String = "&",
        eq: String = "="): String = js.native
  }

  object OSWithNativeVal {
    @js.native
    @JSImport("os", "EOL")
    val EOL: String = js.native

    @js.native
    @JSImport("os", "EOL")
    def EOLAsDef: String = js.native
  }

The rhs of such members must be `= js.native`, and they must have a
native JS load spec annotation (`@JSGlobal` or `@JSImport`).

When accessing such a member, its value is looked up according to
its native load spec, and for methods it is then called as a JS
function.

---

Implementation-wise, these changes require that we can attach a JS
native load spec to *members* of classes, whereas before only
`ClassDef`s could have native load specs. We choose to add a new
kind of `MemberDef` for this purpose:

  JSNativeMemberDef(flags: MemberFlags, name: MethodIdent,
      jsNativeLoadSpec: JSNativeLoadSpec)

along with a term node to read their values:

  SelectJSNativeMember(className: ClassName, member: MethodIdent)

The call sites identify `@js.native` members and, instead of
compiling them to Scala `Apply` nodes, use `SelectJSNativeMember`
(wrapped in a `JSFunctionApply` for methods).

Note that those members have *method* names, even though they are
simple values. We will come back to this below.

Needed native JS members are tracked by the analyzer, using a
dedicated "namespace".

The emitter emits native JS members exactly like existing classes
and modules.

---

Several alternatives were considered and rejected, which led to the
present design. Here is a complete history of ideas:

First, I thought we would make the public binary API use normal
Scala methods, whose *body* would encode, in an implementation-
specific way, the loading of the value and the function call. This
would have allowed native members to implement and override
inherited members. However, the idea was killed by the requirements
for default parameters: it is not possible, inside the body of the
method, to retroactively decide what parameters must be left out or
not.

Once it was clear that the interpretation of the parameters had to
be handled at call site, the problem of *overloading* appeared.
Short of forbidding overloads of JS native defs, which would not
have been consistent with methods in JS types, the binary API had
to preserve JVM overloading semantics. This forced the unintuitive
design that native JS members have *method names*, as opposed to
*field names*, although they behave more like fields.

Another alternative that would have avoided the need for method
names would have been to forbid overloads *that have different
native load specs*. However, that would also have been inconsistent
with methods of JS types (who can have different `@JSName`s), and
would still have required an awkward binary API.

In the end, the design therefore settles on method names for
dedicated node types.

The particular choice of `MethodName` invalidates a previously nice
aspect of the IR, which was that each kind of `Name` was used for
one purpose only. Now, `MethodName`s are used for both Scala
methods and native JS fields. We could have used a different kind
of name, isomorphic to `MethodName` (either by wrapping or
replicating the structure) to avoid that. However, this was deemed
unnecessary.