Fix #3283: Fix incremental linking on configuration changes.
The output of the linker depends on both source code and linker
settings; a change in either should cause linking to happen again.

Previously, the sbt plugin would only do an incremental re-linking when
the source had changed; it would not re-link if only the config had
changed.

This commit adds methods to the linker interface to "fingerprint"
linker configurations.

The sbt plugin writes fingerprints of linker configuration keys to disk
at every run of fastOptJS and fullOptJS. It then compares the
fingerprints of the current configuration objects to those saved to
disk on the previous run in order to determine whether it should re-run
linking.

If both source and linker config have changed, the re-linking only
happens once.