Fix #4034: Encode -x as -1.0 * x instead of 0.0 - x.
The result of `-x` is not the same as `0.0 - x` when `x` is `0.0`:
it is -0.0 instead of +0.0. Therefore, encoding `-x` as `0.0 - x`
was incorrect, as well as link-time optimizations that interpreted
`0.0 - x` as `-x`.

We now encode `-x` as `-1.0 * x` instead, which always give the
correct result. We adapt the rules in the optimizer and in the
emitter to recognize `-1.0 * x` as `-x`, instead of `0.0 - x`.

In addition, we remove some wrong simplifications in the optimizer,
that were assuming that `+` is associative for doubles, whereas it
is not.

All of the above also applies to floats.

Only code compiled with the newer version of the compiler will be
fixed.

Attention! Code doing `-x` in source code and compiled with an
earlier version of the compiler will be *broken* by this change,
since it was encoded as `0.0 - x` in the IR, which used to be
emitted as `-x` (fixing the wrong encoding) and now emitted as
`0.0 - x` (which is correct given the IR, but incorrect given the
original source code). An alternative would be to add a
deserialization hack to deserialize `0.0 - x` as `-1.0 * x`.