Overhaul PrepJSInterop to be simpler, more uniform, with less code paths.
`PrepJSInterop` was a huge bag of spaghetti code, with more checks
added on top of the existing checks as the years passed, without
ever reconsidering the code paths.

This commit is a major overhaul of that phase, which several
advantages.

First, we clearly separate the handling of a) member defs,
b) templates and c) statements/expressions.

Second, checks are structured hierarchically so that each check is
performed at exactly one place, rather than being repeated for
every kind of member def that requires the checks. This makes it
much easier not to forget obscure combinations of features that
would be missing checks. As a "proof" of that claim, each
non-helper method is called from exactly one call site.
(A helper method only inspects some properties of a symbol, as
opposed to performing real checks, reporting errors, mutating the
symbol table and/or transforming the tree.)

Exceptions are:

* `validateJSConstructorOf`, called both from the transformation of
  `js.constructorOf[T]` and that of
  `js.ConstructorTag.materialize[T]`. This could be avoided
  by re-`transform`ing the tree generated by the materialization,
  but that would perform useless work.
* `checkSetterSignature`, called both for setters in JS types and
  exported setters in Scala types.
* `checkRHSCallsJSNative`, called both for concrete members in
  native JS types and for native JS vals and defs.

Some error messages are changed as a direct consequence of the
above changes. No effort is made in this commit to change error
messages beyond those.