Generalize and improve many error messages.
After the major cleanup of the parent commit, we can easily
generalize and improve many error messages reported by
`PrepJSInterop`.

A particular effort is made towards generalizing messages that
would equality apply to several kinds of members (classes and
objects, classes and traits, classes, vals and defs, etc.) instead
of specializing the messages for the particular kind of member on
which the error is reported.

For example, instead of

    Native JS classes must have exactly one
    annotation among @JSGlobal and @JSImport.

we now emit

    Native JS classes, vals and defs must have exactly one
    annotation among @JSGlobal and @JSImport.

A concrete consequence in the code is that the helper method
`kindStrFor()`, and similar helpers, are removed.