Define `unary_?` methods without `()`.
The desugaring of `!x` is `x.unary_!`, not `x.unary_!()`. Since
Scala 2.13.3 now emits warnings when calling a `()`-def without
`()` at call site, we get warnings for `!x` (and similarly `+x`,
`-x` and `~x`).

This commit defines them without `()` to fix this issue.

This is a binary compatible change (both forward and backward) but
it is not backward source compatible for call sites that explicitly
call `x.unary_!()` instead of using the sugar `!x`.