Optimize away top-level `_return: {}` blocks in `void` methods.
For non-void methods, we use `pushLhsInto`, which necessarily
receives the `Lhs.ReturnFromFunction` lhs. It can then pass that on
to a top-level `_return[T]: {}` labeled block.

However, for void methods, this code path was not triggered. We
therefore need to specially handle a top-level `_return: {}` block.
This allows to compile the following IR

  def foo(x: int) {
    _return: {
      if (x > 5)
        return@_return (void 0)
      println(x)
    }
  }

into the following JS:

  foo(x) {
    if (x > 5)
      return (void 0);
    println(x);
  }

instead of

  foo(x) {
    _return: {
      if (x > 5)
        break _return;
      println(x);
    }
  }