Generalize the codegen of LabelDef's for while and do while loops.
Previously, the identification of the shapes of trees produced for
`while` and `do while` loops was done on the scalac Tree, before
transformation. This was problematic because we could *miss* jumps
to the LabelDef buried within the `rhs`, causing assertion errors.

This has never happened in practice before, but is made possible by
the new `async` transformation for Scala 2.12.12+, where `while`
labels can have jumps buried arbitrarily deep inside their rhs.

We fix this issue by doing the general transformation first, and
recording how many jumps are made to the LabelDef. If there are no
more than one, we then safely optimize some shapes, which happen to
cover the shapes of all `while` and `do while` loops, plus some.
We also remove a trailing jump in the rhs even if it is not the
only one---an optimization that is useless for standard loops but
that is triggered with the async transformation.

New `OptimizationTest`s make sure that we cover all shapes of
`while` and `do while` loops. Handling all of them required added a
touch of dead code elimination in `genBlock()`.