Generalize the codegen of `LabelDef`s for pattern matching.
In this commit, we generalize the shapes patmat-generated labels
that we can recognize and translate. Previously, the shape had to
be

  {
    prologue
    case1() {
      ...
    }
    case2() {
      ...
    }
    ...
    case N() {
      ...
    }
    matchEnd(x) { // note: matchEnd is mandatory
      ...
    }
    // note: nothing in the Block after the matchEnd
  }

with three inner shapes for the `matchEnd` itself. Otherwise, the
transformation would bail with an assertion.

The new async implementation coming in Scala 2.12.12 emits new
shapes of patmat labels, and in particular it can produce the
following shape:

  {
    prologue
    case1() {
      ...
    }
    case2() {
      ...
    }
    ...
    case N() {
      ...
    }
    rest
  }

which has no `matchEnd`, and has `rest` code after the final case.
Moreover, it can also generate orphan `matchEnd` labels (without
any `case` preceding them).

We generalize the identification of patmat labels so that we can
recognize the following shapes:

  {
    prologue
    case1() {
      ...
    }
    case2() {
      ...
    }
    ...
    case N() {
      ...
    }
    <optional>:
      matchEnd(x) {
        ...
      }
    rest
  }

including in cases where more sequences are found in `rest`.

Orphan `matchEnd`s automatically fall back on the normal
"enclosing" translation of LabelDefs (and end up having no jumps to
them, which means they are optimized away).