Generalize the codegen for label-apply's.
Previously, we had basically 3 different kinds of labels, each of
which had specific support in `genLabelApply` to translate them in
a way that corresponds to how their `LabelDef`s are handled:

* for patmat `case`s,
* for patmat `matchEnd`s, and
* for any other label working on the basis of surrounding label
  def.

In this commit, we unify all those into a single way to compile
label defs: using the enclosing label def mechanism. Of course,
patmat label defs are not actually enclosing their applys, but we
can rearrange the transformation in advance so that `genLabelApply`
does not need to know about the different kinds.

We store in the dynamic scope (`ScopedVar`s) the necessary info to
generate any given label-apply. Those info are agnostic of the
purpose of the label; only of how their apply's should be
generated.

This allows to decouple the generation of label-apply's from how
their `LabelDef`s are recognized and compiled.