Create test folder for scala 2.x only
The goal is to have a set of tests that are compatible between scala 2.x and dotty.
Since dotty does not has implicit instance for ClassTag[Nothing] and ClassTag[Null].
Tests that require those instances will only be executed on Scala 2:
ArrayBuilderTestScala2 and ClassTagTestScala2 was created to accommodate
test cases from ArrayBuilderTest and ClassTagTest.