Move GlobalVarRef avoidence to the top level of the flow
This was a remnant of when we emitted trees with a mutable builder, so
we had to be careful not to emit trees twice.

We move it to the top-level which:
- Makes conceptually more sense.
- Doesn't require `genAllClasses` to know about the collection
  structure of all classes.

The second point is relevant to module splitting, where classes become
a `Map[Module.ID, List[LinkedClass]]`.

As a side effect, we inline `startRun` and `endRun`. These methods
were not named semantically meaningful (they were named after their
call site, not their function). I triped over it several times, trying
to figure out where the GlobalKnowledge is updated.