Add an IDE stub for `emitter.PrivateLibHolder` in the linkerJS.
In the JVM version of `linker`, `PrivateLibHolder` is a normal
source file, but in JS it is generated by the build. Depending on
whether Metals chose to associate `Emitter.scala` to the JVM or JS
project, we could get spurious compile errors in the IDE because it
would not find `PrivateLibHolder` in the JS project.

We fix this by adding an IDE stub for `PrivateLibHolder` in the JS
version of the linker, similarly to what we do for other generated
sources.