Fix #4061: Report a linking error on export of non-JS ident in a Script.
When emitting a Script (`NoModule`), we cannot export on the
top-level under a name that is not a valid JavaScript identifier.
Such exports must be emitted as `var`s or `let`s, which is only
possible for valid JS identifiers.

Previously, attempts at such invalid exports would silently result
in invalid .js code, producing a `SyntaxError` at run-time. In this
commit, we preemptively report a linking error instead (as a thrown
`LinkingException`).

It can be argued that we might want fallback behaviors for these
cases, as discussed in #4061. However, this commit is a definitive
improvement over the status quo, and leaves the door open for
fallbacks in the future, if they prove desirable/necessary.