Use the published artifacts of scalajs-{logging,js-envs,env-nodejs}.
Now that they are independently published, we simply depend on them
through regular `libraryDependencies`.

The sources of `logging` are moved to the `js/` sources of the
`linker-interface`, because there is no published version of
`scalajs-logging` for Scala.js. We simply keep a copy of the
sources rather than fetching them through the source jars because
it has a very small footprint and is unlikely to change.

The sources of `js-envs`, `js-envs-test-kit` and `env-nodejs` are
deleted.