Fix #4086: Expose `object`s and `class`es in top-level JS objects.
`object`s (resp. `class`es) in top-level objects are static, and
therefore do not receive getters from scalac (resp. the inner JS
class treatment of `ExplicitInnerJS`). This is problematic if the
enclosing top-level object is a JS object, as it prevents
JavaScript code from accessing those objects and classes as members
of the top-level object.

To fix this issue, we extend the `ExplicitInnerJS` phase to:

* Apply the existing transformation of JS classes if they are in a
  static JS object (with an rhs for the field which is just
  `js.constructorOf[Inner.C]`).
* Create exposed getters for objects (Scala and JS) if they are in
  a static JS object.