Add ModuleKind as an attribute to linked Scala.js files
Thanks to this, we do not need to "peek-back" as far when assembling the
jsEnvInput. That in turn allows us to properly deal with different
module kinds in different configurations.

The re-use of the ModuleKind type is somewhat a shortcut, but I think
that is acceptable for the sake of consistency / brievity.