Do not erase a regular class with a static forwarder class.
This is only relevant for case insensitive file systems. On such
file systems, a static nested object could generate a static
forwarder class whose name differed from a regular class' name only
in case, therefore erasing the latter on case insensitive file
systems.

We now avoid emitting static forwarder classes for objects if they
would clash with a regular class on a case insensitive file system.

When doing so, we emit a warning. On Scala 2.13.2+, the warning can
be silenced with `@nowarn("other")` on the object, or
`-Wconf:cat=other:s` in the scalac options.

For top-level objects, the warning is similar to a warning already
emitted by the JVM back-end. For non-top-level objects, it only
happens in Scala.js, and only when they are emitted due to the
opt-in option.

This commit complements its parent in fixing #4148 for good, as now
that issue won't present itself even when we opt in for static
forwards of non-top-level objects. We therefore add a test for
that issue, which we could not do before because the test suite as
a whole requires the opt-in for other reasons.