Fix #4148: Do not emit static forwarders for non-top-level objects by default.
Scala/JVM does not emit static forwarders for non-top-level
objects. The reason we do it in Scala.js is for implementations of
JDK classes that contain static methods in inner static classes
(#3950). However, this causes issues with existing Scala projects
that have non-top-level `object`s and `class`es with names
differing only in case.

In this commit, we now only emit static forwarders for top-level
objects by default, and add an opt-in option to emit them for
non-top-level objects, which should be used by implementations of
JDK classes.