Remove two irrelevant tests that are incompatible with Dotty.
`() + string` and `false + string` are not primitive methods, but
rather use the deprecated `any2stringAdd` extension (unlike for
numeric types and characters). Therefore, the tests for that were
irrelevant since they do not compile a string concatenation where
the lhs is a boolean or an undefined.

Since they are incompatible with Dotty, which treats
`any2stringAdd` as a compile error, we remove those two tests.