Fix #4171: Fix a missing condition in tryOptimizePatternMatch.
`tryOptimizePatternMatch` tries to identify `Return`s to a given
`Labeled` block, and rearrange the code to get rid of both. It was
too liberal in how it did so, as it recognized *any* `Return` in
the expected positions. We now specifically verify that the
`Return`s are for the particular `Labeled` block that we are
optimizing.