Fix #4151: Report a somewhat helpful message for back-end exceptions.
This implements exactly the same last resort mechanism as the JVM
for back-end crashes. In particular, it applies for string literals
that are too long to be serialized (in JVM bytecode or in SJSIR).

Arguably, this should be caught earlier as a nice front-end compile
error, but that is not a job for Scala.js.