Port two new internal methods of `ClassTag` from upstream.
This is a port of the changes to `ClassTag` in
https://github.com/scala/scala/commit/31d6481b8a1322bd1ecfa6d68c49f6c03ab42f19

The new members are `lazy val`s in the upstream library, but we
use `def`s in Scala.js instead. This is because `ClassTag`s are
cached on the JVM but they are stack-allocated on JS.

Although that commit has not reached the 2.13.x branch, we also
apply the changes to the 2.13.x overrides to be future-proof.