Always have an entry point in sbt tests
Not having an entry point at all is a valid use case for the linker,
but it is a corner case. This surfaces in module splitting, where we
will not be able to cover this case in a fully backwards compatible
manner (the linker will not produce anything at all in this case).

We bring the sbt tests (that do not test this corner case) in line
with a more normal use case to ease that transition.