Use captures instead of ctor params for anon JS class captures.
The logic in `GenJSExports.ExportedSymbol` has always separated
"lifted" JS constructors from normal methods and Scala constructors
(the latter are, by construction, never lifted at that point). For
lifted constructors, it separates the original parameters from the
captures, so that orignal parameters become formal parameters of
the JS constructor function, while capture parameters become JS
class captures.

However, before this commit, it put all parameters back together as
regular parameters if the class was an anonymous JS class. The
logic in `genAnonJSClass` reassigned those parameters based on the
actual arguments inserted by lambdalift in the `new` invocation.
Effectively, this dance treated those parameters as class captures,
but with an extra indirection of going through constructor
parameters.

In this commit, we remove this special case in `ExportedSymbol`,
and treat all lifted JS constructors in the same way.
`genAnonJSClass` now passes the actual arguments inserted by
lambdalift as the actual values for the JS class captures thus
produced.

Overall, the end result is basically the same, modulo some boxing
and unboxing which happened with the previous scheme, which the
optimizer would take care of. The new translation is however
cleaner.