Add a javalib-ext-dummies to allow more testing.
We have some rare methods in the core that depend on things that
are not provided by the core. An example is
`java.util.Date.toInstant()`, which depends on the non-provided
`java.time.Instant`. It was previously not possible to test those
methods, which makes them fragile.

In this commit, we introduce a new internal "JDK extension" library
with dummy implementations of those external classes. We start here
with a partial implementation of `java.time.Instant`, which is
enough to test `java.util.Date.toInstant()`.

We add that library in the dependencies of `testSuiteEx` only.
Tests relying on the dummies must go to that test suite, so that
the main test suite is not polluted by the dummies.

The implementation of `java.time.Instant` that we add is also
sufficient to test `j.u.Date.from(Instant)`, but that one is
revealed to be buggy (#4210) and hence we do not add a test for it.