Specify that ROOT is the default locale across the board.
Until now, the default locale was specified as `en-US`. In this
commit, we officially change that default to the root locale,
`Locale.ROOT`.

We make this switch because the root locale is less biased than
en-US.

In practice, this changes nothing, because the root locale and
en-US behave the same way for all the features that have been
implemented so far.