Fix compliance of `String.compareTo` and `{equals,compareTo}IgnoreCase`.
The precise result `String.compareTo` is actually specified, so
that testing only `< 0` or `> 0` is not good enough. We now test
the precise results, and adapt the implementation to conform to
them.

Likewise, the way `equalsIgnoreCase` and `compareToIgnoreCase`
compare characters is specified as a char-by-char normalization.
We fix our implementations to conform to the specified comparison
and adapt our tests accordingly.