Do not rely on hi-IN to get Devanagari digits; use `u-nu-deva`.
While in JDK 8, the locale `hi-IN` would use Devanagari digits, it
is not the case in JDK 11 anymore. It may have been a bug in JDK 8,
although I could not find a bug report mentioning this issue.

Instead, we can get Devanagari digits by adding the Unicode
extension `nu-deva` (i.e., the extension `u-nu-deva`). We fix the
test to do that, and therefore have to enhance our dummy `Locale`
to support a way to create that locale with its extension.