Apply localization when uppercasing non-numbers in `j.u.Formatter`.
Previously, we were never applying locale-specific `toUpperCase()`
in `java.util.Formatter`. This was the result of misinterpreting
the specification, confirmed by an actual bug in JDK 8 as well:
https://bugs.openjdk.java.net/browse/JDK-8060094

The bug was fixed in JDK 11, so in this commit we fix the test and
the implementation. We verify that numbers are still not localized.