Relax a test of `URLDecoder.decode` on the JVM.
The specification says that no exception should be thrown when the
character set need not be consulted. This was respected by JDK 8,
but apparently not by JDK 11 anymore, although there is a very old
bug report that says the not-throwing behavior was as-designed:
https://bugs.openjdk.java.net/browse/JDK-4832991

Therefore, in this commit we are a bit more tolerant on the JVM,
and accept either a correct result or an
`UnsupportedEncodingException`.