Adapt tests about the Unicode Char Database so that they pass on JDK 11.
Newer versions of the JDK use newer versions of the Unicode
database, which invalidate some of our tests. In this commit, we
only adapt the tests so that they pass on JDK 11 as well as JDK 8.
Actually updating our databases to reflect the state of JDK 11 is
left for separate work.