Remove obsolete tests for `js.Dynamic.x` handling.
We used to have a hacky workaround in Scala 2.10 to allow the
selection of members named `x` on `js.Dynamic`. Since we removed
support for 2.10, that workaround was removed, and the tests are
therefore meaningless.

The second test was also the only test in that file to require JS
exports. Removing it makes it easier to validate `js.Dynamic`
without any dependency on exports.