Avoid warnings about `equals` between dangerous types in the test suite.
Scala 2.13.4+ emits warnings when calling `a.equals(b)` if it might
not be consistent with `a == b`, telling to use `a == b` instead.

In a few places in the test suite, we specifically test for the
`equals` behavior, and do not want the `==` behavior. We avoid the
warnings with some no-op rearrangements, or by casting both sides
to `AnyRef` if we cannot do better.