Simplify the doubleToLong conversion when Longs are encoded as BigInts.
The previous implementation was necessary when the `BigInt(d)`
constructor threw for `abs(d) > 2^53`. That throwing behavior was
revised in the BigInt proposal before it reached Stage 4, and
ultimately the behavior specified in ECMAScript is to throw only
for non-integral values.

The conversion still throws for `NaN`, which we have to
special-case, as well as for non-integral values, which is why we
need `Math.trunc`.

The change in the BigInt proposal was in part triggered based on
our feedback, as a matter of fact:
https://github.com/tc39/proposal-bigint/issues/132