Always emit `var`s for internal vars instead of `let`s and `const`s.
Due to their TDZ (Temporal Dead Zone), `let`s and `const`s are much
harder for engines to correctly optimize than `var`s. One recent
(as of this writing) example it the WebKit issue
https://bugs.webkit.org/show_bug.cgi?id=199866.

As the above issue highlights, engines need to perform specific
analyses of the scopes of `let`/`const`s in order to remove the TDZ
checks. For `var`s, there are no TDZ checks to perform to begin
with. Using `var`s is therefore safer, as in more resistant to
unknown performance cliffs.

Since the difference in semantics is not observable from the
outside for internal variables, we can use `var`s all the time.

There is one exception: top-level exports in Script (`NoModule`)
mode are always emitted as `let`s under ECMAScript 2015, since they
are visible from the outside.

A hint can be given to the linker not to avoid `let`s and `const`s
with

    scalaJSLinkerConfig ~= {
      _.withESFeatures(_.withAvoidLetsAndConsts(false))
    }