Do not use ES `class`es for non-Throwable Scala classes.
Benchmarks show that ES `class`es are significantly (up to 10x)
slower than `function`s and prototypes in Firefox.

For semantics reasons, we *must* emit true classes for all
JavaScript classes: they are required to be able to extend true
native `class`es, as well as to inherit static members from parent
classes.

For debugging purposes, it is desirable to emit `Throwable` and
their subclasses as true classes, since that makes them proper
error classes from the JavaScript engine's point of view, which
therefore attach better debugging information to them.

However, there is no requirement to emit other Scala classes as
true ES classes. Since we expect that non-Throwable Scala classes
dominate the uses of other classes, getting good performance even
only for them is beneficial.

In this commit, we revert to emitting `function`s and prototypes
for all Scala classes that do not extend `Throwable`, even when
emitting in ECMAScript 2015 mode.

This brings benchmark results for Firefox in ES 2015 basically on
par with ES 5.1 as well as with Chrome (with minor variations).

The only downside to this change is code size, which significantly
increases. However, it is probably a relatively small price to pay
not to have 10x performance regressions on Firefox.

It is possible to hint to the linker to use `class`es for all
classes with

    scalaJSLinkerConfig ~= {
      _.withESFeatures(_.withAvoidClasses(false))
    }