Build support for Scala 3 in sbt-scalajs.
Currently, sbt-dotty knows about Scala.js, and appears to always be
applied *after* sbt-scalajs, which allows it to fix up the
configuration for Scala 3.

However, most of the functionality of sbt-dotty will have to be
merged in sbt itself, where we cannot bake Scala.js support.

This commit adds direct support for Scala 3 in sbt-scalajs, in a
way that works today (if sbt-dotty is present) as well as
future-proofs for when sbt-dotty won't be used anymore.